// client/src/shared/enums/Abilities.ts
export enum Abilities {
  Club = '1',
  Funding1 = '2',
  Funding2 = '3',
  InPosition1 = '4',
  InPosition2 = '5',
  InPosition3 = '6',
  InPosition4 = '7',
  InPositionLast = '8',
  MaxWordLength4 = '9',
  MaxWordLength5 = '10',
  MaxWordLength6 = '11',
  MaxWordLength7 = '12',
  MinWordLength4 = '13',
  MinWordLength5 = '14',
  MinWordLength6 = '15',
  MinWordLength7 = '16',
  NextToVowel = '17',
  NextToWild = '18',
  NotNextToVowel = '20',
  NotNextToWild = '19',
  OtherInPosition1 = '21',
  OtherInPosition2 = '22',
  OtherInPosition3 = '23',
  OtherInPosition4 = '24',
  OtherInPosition5 = '25',
  Retain = '26',
  RetainLeft = '27',
  RetainRight = '28',
  Vowels = '29',
  Wilds = '30',
  WordLength4 = '31',
  WordLength5 = '32',
  WordLength6 = '33',
  WordLength7 = '34',
  Wild = '35',
  CopyAbilityInPosition1 = '36',
  CopyAbilityInPosition2 = '37',
  CopyAbilityInPosition3 = '38',
  CopyAbilityInPosition4 = '39',
  CopyAbilityInPosition5 = '40',
}